requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Deprecated==1.2.18
dotenv==0.9.9
google==3.0.0
google-auth==2.40.3
google-genai==1.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.2
